% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagnn-package.R
\docType{package}
\name{dagnn-package}
\alias{dagnn}
\alias{dagnn-package}
\title{dagnn: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rdinnager.github.io/dagnn/}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
